requirements.txt
faiss-cpu==1.6.1
h5py==2.10.0
numpy==1.18.1
Pillow==6.2.1
scikit-learn==0.22
torchvision==0.4.2
tqdm==4.41.1
